duplicate and rename logging function to allow logan and quadra to be used together
