+++ xcoder/logging_shim.c
@@ -10,14 +10,22 @@
 #include <ni_util.h>
 #endif
 
+#ifdef LOGAN
+void rust_netint_logan_callback(int level, char* message);
+#else
 void rust_netint_callback(int level, char* message);
+#endif
 
 void netint_log_callback(int level, const char* format, va_list args) {
   char buf[2048] = {0};
   size_t buf_len = sizeof(buf) / sizeof(buf[0]);
   int chars_written = snprintf(buf, buf_len, format, args);
   if (chars_written > -1 && chars_written + 1 < ((int) buf_len)) {
+      #ifdef LOGAN
+      rust_netint_logan_callback(level, buf);
+      #else
       rust_netint_callback(level, buf);
+      #endif
   }
 }
